+++ core/components/molecules/form/field/stories/field.story.js
@@ -14,7 +14,7 @@ import {
   Button
 } from '@auth0/cosmos'
 
-storiesOf('Form').add('all the fields', () => (
+storiesOf('Form', module).add('all the fields', () => (
   <Example title="all the fields">
     <Form>
       <Form.Field label="Field label">
@@ -29,12 +29,14 @@ storiesOf('Form').add('all the fields', () => (
       </Form.Field>
       <Form.Field label="Options list">
         <Select
+          value={1}
           options={[
-            { text: 'First option', value: '1', defaultSelected: true },
+            { text: 'First option', value: '1' },
             { text: 'Second option', value: '2' },
             { text: 'Third option', value: '3' },
             { text: 'Fourth option', value: '4' }
           ]}
+          onChange={event => console.log(event)}
         />
       </Form.Field>
       <Form.Field label="Subscribe">
@@ -56,14 +58,16 @@ storiesOf('Form').add('all the fields', () => (
           </Checkbox>
         </Checkbox.Group>
       </Form.Field>
-      <Form.Field label="Custom field">
-        <input type="date" />
+      <Form.Field htmlFor="bring-your-own-id" label="Custom field">
+        <input id="bring-your-own-id" type="date" />
       </Form.Field>
       <Form.Field label="Height" helpText="How tall are you?" error="Show only in the first field">
         <Stack>
           <TextInput placeholder="Value" hasError />
           <Select
+            value="cm"
             options={[{ text: 'centimetres', value: 'cm' }, { text: 'inches', value: 'in' }]}
+            onChange={event => console.log(event)}
           />
           <Button appearance="link" icon="copy" label="Copy value" onClick={e => console.log(e)} />
         </Stack>