+++ changelog.md
@@ -4,6 +4,82 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/). Currently, this project is using an `0.x` versioning system to indicate its preview status. As such, it does _not_ currently adhere to [Semantic Versioning](http://semver.org/spec/v2.0.0.html). Once we reach production readiness, we will follow semver beginning with our `1.0` release. In the meantime, please read this changelog for information on breaking changes!
 
+## 0.19.0 [February 26, 2019]
+
+### Changed
+
+- Select: Use `react-select` wrapper only on `multiple` and/or `searchable` selects. [#1528]
+
+### Fixed
+
+- Layouts: avoid rendering empty children. [#1529]
+- Select: Fix dialog overflow when using a `react-select` based Select. [#1527]
+- TextArea: Add a space when using actions in Textarea. [#1523]
+
+## 0.18.0 [February 22, 2019]
+
+### Added
+
+- Select: `react-select` wrapper. [#1512]
+- Tooltip: new variant for input action tooltips. [#1522]
+
+## 0.17.0 [February 19, 2019]
+
+### Added
+
+- Docs: Display Layouts section [#1510][#1511]
+- PageHeader: Support passing React nodes as `title`. [#1518].
+- Avatar: Add image resolution failover. [#1515]
+- Docs: New authoring guide. [#1513]
+- Docs: Document build pipeline. [#1514]
+
+### Fixed
+
+- Form.Field: Automatically add id to fields. [#1439]
+- Checkbox: Improve `readOnly` visual state. [#1517]
+
+### Changed
+
+- Image: Rename `source` to `src` (via soft deprecation) [#1505]
+- Switch: Rename `onToggle` to `onChange` (via soft deprecation) [#1506]
+
+## 0.16.2 [February 13, 2019]
+
+### Fixed
+
+- Tooltip: Use Popper.js' position fixed mode. [#1501]
+- Sidebar: Fix styles not changing when selected link prop changed. [#1500]
+
+## 0.16.1 [February 7, 2019]
+
+### Fixed
+
+- Sidebar: Fix sidebar linkgroup item colored when open but no subitems selected [#1498]
+- Tooltip: Apply z-index back to Tooltip Element [#1497]
+- Dialog: Revert Composable actions for Tabs in Dialog [#1496]
+- Icon: Fixed typo in icon ‘brand-weibo’ [#1495]
+
+## 0.16.0 [February 7, 2019]
+
+### Added
+
+- Docs: Start migration guide [#1490]
+- Dialog: Composable actions for Dialog + Tabs [#1458]
+
+### Fixed
+
+- Sidebar: Fixed selected color for Link group and added basic stories [#1493]
+- Select: Allow using empty string as a value [#1483]
+- Checkbox: Fix onClick event firing twice [#1473]
+- ResourceList: Add check for onItemClick [#1491]
+- ResourceList: Fix Item body width validator [#1481][#1482]
+- Layouts: Update gutter sizes prop-type [#1480]
+- Table: Fix story image source [#1479]
+
+### Internal
+
+- Remove sketch files [#1485]
+
 ## 0.15.1 [January 31, 2019]
 
 ### Fixed